DDA-69: Rename room to diograph
