+++ src/react/features/diograph/actionsTypes.js
