+++ src/react/features/diograph/hooks.js
@@ -1,14 +1,14 @@
 import { useStore } from '../../store'
 
 export const useDiorys = (ids = {}) => {
-  const [{ diograph = {} }] = useStore((state) => state.room)
+  const [{ diograph = {} }] = useStore((state) => state.diograph)
   return Object.entries(ids)
     .map(([key, { id }]) => ({ key, ...diograph[id] }))
     .filter(({ id }) => id)
 }
 
 export const useDiory = (id) => {
-  const [{ diograph = {} }] = useStore((state) => state.room)
+  const [{ diograph = {} }] = useStore((state) => state.diograph)
   const diory = diograph[id]
   const links = diory && diory.links
   return {
