+++ src/react/features/filters/hooks/useTextFilter.js
@@ -1,13 +1,13 @@
 import { useDispatch, useStore } from '../../../store'
 
 import { setFocus } from '../../navigation/actions'
-import { createDiory } from '../../room/actions'
+import { createDiory } from '../../diograph/actions'
 import { setTextFilter } from '../actions'
 
 import { TEXT_FILTER_DIORY_ID } from './textFilterDioryId'
 
 export const useTextFilter = () => {
-  const [{ diograph }] = useStore((state) => state.room)
+  const [{ diograph }] = useStore((state) => state.diograph)
   const [{ focus }] = useStore((state) => state.navigation)
   const [{ textFilter }] = useStore((state) => state.filters)
