+++ src/react/features/lenses/useLensActions.js
@@ -1,7 +1,7 @@
 import { useDispatchActions } from '../../store'
 import { setActive } from '../buttons/actions'
 import { setFocus } from '../navigation/actions'
-import { createDiory, createLink, deleteDiory, deleteLink, updateDiory } from '../room/actions'
+import { createDiory, createLink, deleteDiory, deleteLink, updateDiory } from '../diograph/actions'
 
 export const useLensActions = () => {
   const { dispatchAction } = useDispatchActions()
