+++ src/react/features/navigation/hooks/useDocumentTitle.js
@@ -2,7 +2,7 @@ import { useEffect } from 'react'
 import { useStore } from '../../../store'
 
 export const useDocumentTitle = () => {
-  const [{ diograph = {} }] = useStore((state) => state.room)
+  const [{ diograph = {} }] = useStore((state) => state.diograph)
   const [{ path }] = useStore((state) => state.navigation)
   useEffect(() => {
     document.title = ['Home']
