+++ src/react/features/tools/create/CreateTool.js
@@ -2,9 +2,9 @@ import React from 'react'
 import { v4 as uuid } from 'uuid'
 import { useDispatchActions, useStore } from '../../../store'
 import { useButtons } from '../../buttons'
-import { useFocusDiory } from '../../room/hooks'
+import { useFocusDiory } from '../../diograph/hooks'
 
-import { createDiory, createLink } from '../../room/actions'
+import { createDiory, createLink } from '../../diograph/actions'
 
 import UpdateView from '../update/UpdateView'
