+++ src/react/store/initialState.js
@@ -4,7 +4,7 @@ import { initialState as filters } from '../features/filters/reducers'
 import { initialState as home } from '../features/home/reducers'
 import { initialState as lenses } from '../features/lenses/reducers'
 import { initialState as navigation } from '../features/navigation/reducers'
-import { initialState as room } from '../features/room/reducers'
+import { initialState as diograph } from '../features/diograph/reducers'
 
 export const initialState = {
   buttons,
@@ -13,5 +13,5 @@ export const initialState = {
   home,
   lenses,
   navigation,
-  room,
+  diograph,
 }
